"Rutas API actualizadas, validación y middleware para estudiantes, mentores y funcionalidades de inicio de sesión"

+++ src/database/serverAuth.ts
@@ -1,31 +1,33 @@
-"use server"
+"use server";
 
 import { JWTValidate } from "@/lib/server/auth";
 import { createResponse } from "@/lib/utils";
 import { sql } from "@vercel/postgres";
 
-export async function ValidateIDSession(session:string) {
+export async function ValidateIDSession(value: string) {
 
-    try {
-        
-        const {rows} = await  sql` SELECT * FROM sesiones WHERE id = ${session} `;
+  const session = value.split(";")[0];
 
-        const validateJWT = await JWTValidate(rows[0].token);
+  console.log(session)
 
-        if(!validateJWT) {
-            await sql` DELETE FROM sesiones WHERE id = ${session} `
-            return createResponse(false, [], "JWT INVALIDO");
-        }
+  try {
+    const { rows } = await sql` SELECT * FROM sesiones WHERE id = ${session} `;
 
-        return createResponse(true, validateJWT, "JWT VALIDO")
-        
-    } catch (error) {
-        
-        console.log(error);
-
-        return createResponse(false, [], "error en la base de datos");
+    if (!rows.length) {
+      return createResponse(false, [], "JWT INVALIDO");
     }
 
+    const validateJWT = await JWTValidate(rows[0].token);
+
+    if (!validateJWT) {
+      await sql` DELETE FROM sesiones WHERE id = ${session} `;
+      return createResponse(false, [], "JWT INVALIDO");
+    }
+
+    return createResponse(true, validateJWT, "JWT VALIDO");
+  } catch (error) {
+    console.log(error);
+
+    return createResponse(false, [], "error en la base de datos");
+  }
 }
-
-
